+++ client/src/app/modules/image-viewers/widgets/context-image/context-image-exporter.ts
@@ -3,7 +3,6 @@ import { ContextImageModel } from "src/app/modules/image-viewers/image-viewers.m
 import { ContextImageDrawModel } from "src/app/modules/image-viewers/widgets/context-image/context-image-model";
 import { SnackbarService, WidgetKeyItem } from "src/app/modules/pixlisecore/pixlisecore.module";
 import { APIEndpointsService } from "src/app/modules/pixlisecore/services/apiendpoints.service";
-import { RGBUPlotModel } from "src/app/modules/scatterplots/widgets/rgbu-plot-widget/rgbu-plot-model";
 import { CanvasDrawer, CanvasParams } from "src/app/modules/widget/components/interactive-canvas/interactive-canvas.component";
 import { PanZoom } from "src/app/modules/widget/components/interactive-canvas/pan-zoom";
 import {