+++ CHANGELOG.md
@@ -1,16 +1,21 @@
-## 4.22.0 (Unreleased)
+## 4.22.0 (2024-06-27)
 
 ### Features
 
 - **Dataset Editing** Adds the ability to edit dataset metadata, including name, description, and tags from the dataset listing page.
   - Shows a markdown preview of the description field.
 - **Chart Dataset Colour Support** Adds support for custom colours for dataset all points in the charts. This feature has existed under the scan configuration tab for a while, but was not previously supported by all charts.
+- **Quant Tracker is back!** It can now be found as a tab in the "Run PIQUANT" screen (where you view the periodic table). Look for the "Quant Jobs" tab. It shows the state of quants you're running and allows viewing or downloading the logs (all PIQUANT logs in one zip file!). You can also view the raw PIQUANT CSV output, and set the quant as the one used in your workspace.
+- **Image Picker shows PMCs** So for uploading images to PIXLISE, you can now see the PMC associated with an image when picking which one you're "matching" your image to.
+- **Beam location versioning improved** PIXLISE had the ability to store multiple version of beam locations already but some updates will now make it work correctly should we run a Beam Geometry Tool v3. V1 beam locations will now be able to be imported into PIXLISE and we'll provide a UI setting to allow switching between them in the next release.
 
 ### Bug Fixes
 
 - Variogram custom expression responses were being treated as invalid due to another recent bug fix. This has been resolved.
 - Fixes bug where variogram doesn't automatically live update when built-in comparison expressions are switched between.
 - Shows warning when selecting quantified elements tab in expression picker when no quantification is configured.
+- Quantified Elements are now shown in atomic number order on the expression picker dialog
+- Made date/time displays a bit more uniform across the application
 
 ## 4.21.0 (2024-06-12)